# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRW REST API",
    "version": "0.33.0",
    "description": "Firecrawl-compatible web scraping, crawling, mapping, search, and extraction API. Hosted at https://api.fastcrw.com; self-host with `docker run -p 3000:3000 ghcr.io/us/crw:latest`.",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://opensource.org/licenses/AGPL-3.0"
    },
    "contact": {
      "url": "https://github.com/us/crw"
    }
  },
  "servers": [
    {
      "url": "https://api.fastcrw.com",
      "description": "Hosted (requires Bearer token)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Self-hosted (no auth by default)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {}
  ],
...
  "paths": {
...
    "/v1/scrape": {
      "post": {
        "tags": [
          "scrape"
        ],
        "summary": "Scrape one URL",
        "operationId": "scrape",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scrape result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/crawl/{id}": {
      "get": {
        "tags": [
          "crawl"
        ],
        "summary": "Get crawl job status and results",
        "operationId": "crawlStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Crawl job id returned by POST /v1/crawl"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status + collected pages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlStatus"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/batch/scrape": {
      "post": {
        "tags": [
          "batch"
        ],
        "summary": "Start an async batch scrape over an explicit URL list",
        "operationId": "batchScrapeStart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch job accepted; poll with /v1/batch/scrape/{id}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchScrapeAccepted"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batch/scrape/{id}": {
      "get": {
        "tags": [
          "batch"
        ],
        "summary": "Get batch job status and results",
        "operationId": "batchScrapeStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Batch job id returned by POST /v1/batch/scrape"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status + scraped pages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlStatus"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "batch"
        ],
        "summary": "Cancel an active batch job",
        "operationId": "batchScrapeCancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Batch job id returned by POST /v1/batch/scrape"
          }
        ],
        "responses": {
          "200": {
            "description": "Job cancelled; status becomes \"cancelled\""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Hosted only. Get a key at https://fastcrw.com/register."
      }
    },
    "schemas": {
...
      "ScrapeRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "formats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutputFormat"
            },
            "default": [
              "markdown"
            ]
          },
          "onlyMainContent": {
            "type": "boolean",
            "default": true,
            "description": "Strip navigation, ads, and boilerplate; keep main article content only"
          },
          "renderJs": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "null = auto-detect, true = force JS rendering, false = skip JS rendering"
          },
          "waitFor": {
            "type": "integer",
            "minimum": 0,
            "description": "Milliseconds to wait after JS rendering before extracting content"
          },
          "includeTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [],
            "description": "CSS selectors / HTML tags to include in extraction"
          },
          "excludeTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [],
            "description": "CSS selectors / HTML tags to remove before extraction"
          },
          "jsonSchema": {
            "type": "object",
            "additionalProperties": true,
            "description": "JSON Schema describing the structured output shape for the 'json' format"
          },
          "basis": {
            "type": "boolean",
            "default": false,
            "description": "Return per-field evidence alongside the 'json' format: each top-level scalar schema property comes back as a Basis with an honest status. Requires a JSON schema; off by default, and when off the request sent to the model is unchanged."
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom HTTP request headers forwarded to the target URL"
          },
          "cssSelector": {
            "type": "string",
            "description": "CSS selector to narrow content extraction to a specific element"
          },
          "xpath": {
            "type": "string",
            "description": "XPath expression to narrow content extraction to a specific element"
          },
          "chunkStrategy": {
            "$ref": "#/components/schemas/ChunkStrategy",
            "description": "Strategy for chunking the extracted markdown"
          },
          "query": {
            "type": "string",
            "description": "Query string for BM25/cosine chunk filtering (used with chunkStrategy)"
          },
          "filterMode": {
            "type": "string",
            "enum": [
              "bm25",
              "cosine"
            ],
            "description": "Filtering algorithm to rank chunks against query"
          },
          "topK": {
            "type": "integer",
            "minimum": 1,
            "description": "Number of top chunks to return after filtering (default: 5)"
          },
          "proxy": {
            "type": "string",
            "description": "Per-request proxy URL (http/https/socks5). Overrides global config."
          },
          "proxyList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [],
            "description": "Per-request proxy pool to rotate among (BYOP). Takes precedence over proxy and server config."
          },
          "proxyRotation": {
            "type": "string",
            "enum": [
              "round_robin",
              "random",
              "sticky_per_host"
            ],
            "description": "Rotation strategy for proxyList (default: sticky_per_host)"
          },
          "country": {
            "type": "string",
            "description": "2-letter ISO 3166-1 alpha-2 country code for residential proxy egress (e.g. 'us', 'gb')"
          },
          "userId": {
            "type": "string",
            "description": "Optional sticky user identity for identifying user sessions. Must match ^[A-Za-z0-9._-]+$ (1-128 chars); blank = no sticky identity."
          },
          "sessionId": {
            "type": "string",
            "description": "Optional sticky session within userId. Only takes effect together with userId; anything else is treated as no sticky identity. Same ^[A-Za-z0-9._-]+$ rule."
          },
          "stealth": {
            "type": "boolean",
            "description": "Override stealth mode for this request (null = use global config)"
          },
          "actions": {
            "type": "object",
            "additionalProperties": true,
            "description": "Unsupported Firecrawl parameter — returns a clear error if supplied"
          },
          "extract": {
            "$ref": "#/components/schemas/ExtractOptions",
            "description": "Firecrawl-compatible extract object (e.g. { schema: {...} })"
          },
          "llmApiKey": {
            "type": "string",
            "description": "BYOK: LLM provider API key for structured extraction (overrides server key)"
          },
          "llmProvider": {
            "type": "string",
            "enum": [
              "anthropic",
              "openai",
              "deepseek",
              "azure",
              "openai-compatible",
              "openai-responses"
            ],
            "description": "Per-request LLM provider override"
          },
          "llmModel": {
            "type": "string",
            "description": "Per-request LLM model identifier override"
          },
          "baseUrl": {
            "type": "string",
            "description": "Per-request LLM base URL override for Chat Completions-compatible or Responses-compatible providers (e.g. https://api.deepseek.com/v1)"
          },
          "summaryPrompt": {
            "type": "string",
            "maxLength": 500,
            "description": "User-supplied instructions appended to the summary system prompt (capped at 500 chars)"
          },
          "maxContentChars": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum bytes of scraped content sent to LLM for 'summary' format (capped at 200 KB)"
          },
          "renderer": {
            "type": "string",
            "enum": [
              "auto",
              "lightpanda",
              "chrome",
              "chrome_proxy",
              "playwright",
              "camoufox"
            ],
            "description": "Pin this request to a specific renderer. Non-auto implies renderJs:true unless renderJs:false is set."
          },
          "deadlineMs": {
            "type": "integer",
            "minimum": 1,
            "maximum": 60000,
            "description": "End-to-end deadline budget in milliseconds (default: 8000, max: 60000)"
          },
          "debug": {
            "type": "boolean",
            "description": "When true, response includes debugExtraction trace with every candidate the extractor considered"
          },
          "changeTracking": {
            "$ref": "#/components/schemas/ChangeTrackingOptions",
            "description": "Activated when formats includes 'changeTracking'. The format entry is the plain string; options ride on this sibling field."
          },
          "goal": {
            "type": "string",
            "maxLength": 2048,
            "description": "Plain-language monitor goal for the meaningful-change judge"
          },
          "judgeEnabled": {
            "type": "boolean",
            "description": "Run the LLM meaningful-change judge on a changed page (requires goal)"
          },
          "parsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParserSpec"
            },
            "description": "Document parser directives. Omit = auto-parse PDFs. Empty array = disable parsing."
          },
          "screenshotFullPage": {
            "type": "boolean",
            "default": false,
            "description": "Capture the whole page rather than the viewport, for `formats: [\"screenshot\"]`."
          }
        }
      },
      "ScrapeResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "warning": {
            "type": "string",
            "description": "Top-level advisory (e.g. fallback render notice)"
          },
          "data": {
            "type": "object",
            "properties": {
              "markdown": {
                "type": "string"
              },
              "html": {
                "type": "string"
              },
              "rawHtml": {
                "type": "string"
              },
              "links": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "json": {
                "type": "object",
                "additionalProperties": true
              },
              "metadata": {
                "$ref": "#/components/schemas/PageMetadata"
              },
              "warning": {
                "type": "string"
              },
              "warnings": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "renderDecision": {
                "type": "object",
                "description": "How the renderer was chosen. A tagged object: `kind` is one of `userPinned`, `autoDefault`, `autoPromoted`, `breakerSkipped`, `failover`, and the remaining fields depend on it.",
                "properties": {
                  "kind": {
                    "type": "string"
                  }
                }
              },
              "creditCost": {
                "type": "number"
              },
              "basis": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Basis"
                },
                "description": "Per-field evidence for the 'json' format; present only when the request set basis:true."
              },
              "basisWarnings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BasisWarning"
                }
              },
              "llmInputHash": {
                "type": "string",
                "description": "'sha256:'-prefixed hash of the canonical text sent to the extraction LLM. The independent record a consumer checks a citation's sourceHash against, so the check is not circular."
              },
              "sourceHash": {
                "type": "string",
                "description": "Fingerprint of the canonical markdown, for dedup and caching."
              },
              "plainText": {
                "type": "string",
                "description": "Plain-text view; present when `formats` includes `plainText`."
              },
              "images": {
                "$ref": "#/components/schemas/ScrapedImages"
              },
              "summary": {
                "type": "string",
                "description": "LLM summary; present when `formats` includes `summary`."
              },
              "llmUsage": {
                "type": "object",
                "description": "Token usage for the LLM call, when one ran."
              },
              "chunks": {
                "type": "array",
                "items": {
                  "type": "object"
                },
                "description": "Chunked content, when a `chunkStrategy` was requested."
              },
              "contentType": {
                "type": "string",
                "description": "MIME type of the fetched resource."
              },
              "changeTracking": {
                "type": "object",
                "description": "Diff result; present when `formats` includes `changeTracking`."
              },
              "screenshot": {
                "type": "string",
                "description": "PNG capture as a `data:image/png;base64,...` URL; present when `formats` includes `screenshot`."
              },
              "block": {
                "type": "object",
                "description": "Anti-bot verdict. Present only when the page was a block or challenge shell, in which case `success` is false."
              },
              "debugExtraction": {
                "type": "object",
                "description": "Extraction trace; present only when `debug: true`."
              }
            }
          }
        }
      },
...
      "CrawlStatus": {
        "type": "object",
        "required": [
          "success",
          "status",
          "total",
          "completed",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "scraping",
              "completed",
              "failed",
              "cancelled"
            ]
          },
          "completed": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScrapeResponse"
            }
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "BatchScrapeRequest": {
        "type": "object",
        "description": "Every `/v1/scrape` body field except `url` is accepted here and applied to each URL. See `ScrapeRequest` for the full set.",
        "required": [
          "urls"
        ],
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "minItems": 1,
            "description": "URLs to scrape. Each is SSRF-validated; invalid entries are reported in invalidUrls."
          },
          "ignoreInvalidUrls": {
            "type": "boolean",
            "default": true,
            "description": "When false, a single invalid URL rejects the whole submit with 400."
          },
          "maxConcurrency": {
            "type": "integer",
            "minimum": 1,
            "description": "How many URLs the job scrapes concurrently. Clamped server-side to crawler.max_batch_concurrency."
          },
          "formats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutputFormat"
            },
            "default": [
              "markdown"
            ]
          },
          "onlyMainContent": {
            "type": "boolean",
            "default": true
          },
          "userId": {
            "type": "string",
            "description": "Optional sticky user identity for identifying user sessions. Must match ^[A-Za-z0-9._-]+$ (1-128 chars); blank = no sticky identity."
          },
          "sessionId": {
            "type": "string",
            "description": "Optional sticky session within userId. Only takes effect together with userId; anything else is treated as no sticky identity. Same ^[A-Za-z0-9._-]+$ rule."
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid bearer token (hosted only)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}